# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List requests
      parameters:
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmpRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmpRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmpRequest"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Approve or reject a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestBody"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmpRequest"

  # ─── PAYROLL ─────────────────────────────────────────────────────────────────
...
  /reports/attendance:
    get:
      operationId: getAttendanceReport
      tags: [reports]
      summary: Attendance report by department and date range
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
  /reports/leave:
    get:
      operationId: getLeaveReport
      tags: [reports]
      summary: Leave utilization report
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Leave report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveReport"
  /reports/payroll:
    get:
      operationId: getPayrollReport
      tags: [reports]
      summary: Payroll report by period
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollReport"

  # ─── SETTINGS ────────────────────────────────────────────────────────────────
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get company settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update company settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"

  # ─── AUDIT LOG ───────────────────────────────────────────────────────────────
  /audit:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit logs (Super Admin only)
      parameters:
        - name: actorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: entityType
          in: query
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    EmpRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userFullName:
          type: ["string", "null"]
        departmentId:
          type: integer
        departmentName:
          type: ["string", "null"]
        requestType:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        reviewedBy:
          type: ["integer", "null"]
        reviewerName:
          type: ["string", "null"]
        reviewerRemarks:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, departmentId, requestType, title, description, status, createdAt, updatedAt]
    CreateRequestBody:
      type: object
      properties:
        requestType:
          type: string
        title:
          type: string
        description:
          type: string
      required: [requestType, title, description]
    ReviewRequestBody:
      type: object
      properties:
        status:
          type: string
        reviewerRemarks:
          type: ["string", "null"]
      required: [status]

    # ─── PAYROLL ───────────────────────────────────────────────────────────────
...
    AttendanceReport:
      type: object
      properties:
        totalPresent:
          type: integer
        totalAbsent:
          type: integer
        totalLate:
          type: integer
        byDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DeptAttendanceStat"
      required: [totalPresent, totalAbsent, totalLate, byDepartment]
    DeptAttendanceStat:
      type: object
      properties:
        departmentName:
          type: string
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
      required: [departmentName, present, absent, late]
    LeaveReport:
      type: object
      properties:
        totalApplications:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
        byType:
          type: array
          items:
            $ref: "#/components/schemas/LeaveTypeStat"
      required: [totalApplications, approved, rejected, pending, byType]
    LeaveTypeStat:
      type: object
      properties:
        leaveType:
          type: string
        count:
          type: integer
        totalDays:
          type: integer
      required: [leaveType, count, totalDays]
    PayrollReport:
      type: object
      properties:
        totalNetSalary:
          type: number
        totalBaseSalary:
          type: number
        totalBonuses:
          type: number
        totalDeductions:
          type: number
        employeeCount:
          type: integer
        approvedCount:
          type: integer
      required: [totalNetSalary, totalBaseSalary, totalBonuses, totalDeductions, employeeCount, approvedCount]

    # ─── SETTINGS ──────────────────────────────────────────────────────────────
    CompanySettings:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        logoUrl:
          type: ["string", "null"]
        timezone:
          type: string
        workStartTime:
          type: string
        workEndTime:
          type: string
        payrollCycleDay:
          type: integer
        lateThresholdMinutes:
          type: integer
        updatedAt:
          type: string
      required: [id, companyName, timezone, workStartTime, workEndTime, payrollCycleDay, lateThresholdMinutes, updatedAt]
    UpdateSettingsBody:
      type: object
      properties:
        companyName:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        workStartTime:
          type: ["string", "null"]
        workEndTime:
          type: ["string", "null"]
        payrollCycleDay:
          type: ["integer", "null"]
        lateThresholdMinutes:
          type: ["integer", "null"]

    # ─── AUDIT ─────────────────────────────────────────────────────────────────
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        actorId:
          type: integer
        actorName:
          type: ["string", "null"]
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, actorId, action, entityType, createdAt]