# lib/api-zod/src/generated/types/empRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface EmpRequest {
  id: number;
  userId: number;
  /** @nullable */
  userFullName?: string | null;
  departmentId: number;
  /** @nullable */
  departmentName?: string | null;
  requestType: string;
  title: string;
  description: string;
  status: string;
  /** @nullable */
  reviewedBy?: number | null;
  /** @nullable */
  reviewerName?: string | null;
  /** @nullable */
  reviewerRemarks?: string | null;
  createdAt: string;
  updatedAt: string;
}