# lib/api-zod/src/generated/types/reviewRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewRequestBody {
  status: string;
  /** @nullable */
  reviewerRemarks?: string | null;
}